requirements.txt
# Network Security Scanner - Installation Guide
# ===========================================

# Installation Steps:
# 1. Create and activate virtual environment:
#    python -m venv venv
#    source venv/bin/activate  # Linux/Mac
#    venv\Scripts\activate     # Windows
#
# 2. Install all dependencies:
#    pip install -r requirements.txt
#
# 3. Install system dependencies:
#    Ubuntu/Debian: sudo apt-get install nmap mysql-server
#    CentOS/RHEL:  sudo yum install nmap mysql-server
#    Windows:      Download and install Nmap from nmap.org
#                 Download and install MySQL from mysql.com
#
# 4. Database setup:
#    mysql -u root -p
#    CREATE DATABASE network_scanner;
#    USE network_scanner;
#    source schema.sql;
#
# 5. Configure environment:
#    Copy .env.example to .env
#    Update database credentials in .env

# Core Dependencies
# ================
streamlit==1.31.1                    # Web framework
extra-streamlit-components==0.1.60   # Additional Streamlit components

# Database
# ========
mysql-connector-python==8.3.0        # MySQL database connector
python-dotenv==1.0.1                 # Environment variable management
SQLAlchemy==2.0.25                   # SQL toolkit and ORM

# Network Scanning
# ===============
python-nmap==0.7.1                   # Nmap scanning interface
scapy==2.5.0                         # Packet manipulation
netaddr==0.9.0                       # Network address manipulation

# Data Processing & Visualization
# =============================
pandas==2.2.0                        # Data manipulation and analysis
plotly==5.18.0                       # Interactive visualizations
numpy==1.26.3                        # Numerical computations

# Security
# ========
bcrypt==4.1.2                        # Password hashing
PyJWT==2.8.1                         # JSON Web Tokens
cryptography==42.0.2                 # Cryptographic recipes
pyOpenSSL==24.0.0                    # TLS/SSL wrapper

# Date/Time Handling
# =================
python-dateutil==2.8.2               # Date utilities
pytz==2024.1                         # Timezone calculations

# Networking & System
# ==================
requests==2.31.0                     # HTTP library
ipaddress==1.0.23                    # IP address manipulation
psutil==5.9.8                        # System and process utilities

# Development Tools
# ================
pytest==8.0.1                        # Testing framework
black==24.1.1                        # Code formatter
flake8==7.0.0                        # Code linter
pylint==3.0.3                        # Code analysis
mypy==1.8.0                          # Static type checker

# Optional Dependencies
# ===================
rich==13.7.0                         # Terminal formatting
tqdm==4.66.1                         # Progress bars
colorama==0.4.6                      # Colored terminal output

# Notes:
# - Versions are pinned for reproducibility
# - Some packages may require system-level dependencies
# - Use pip-compile to generate requirements.txt from setup.py
# - Use pip-sync to ensure exact package versions

# Development Installation:
# pip install -r requirements.txt[dev]
# This will install additional development tools

# Production Installation:
# pip install -r requirements.txt
# This will install only required packages

python-docx==1.0.1
openpyxl==3.1.2
python-whois==0.7.3 